--- Client/vite-project/src/components/admin/RoomDashboard/CardDisable.css
@@ -1,68 +1,0 @@
-.Disable-grid {
-  display: grid;
-  grid-template-columns: repeat(auto-fill, minmax(250px, 1fr));
-  gap: 40px;  
-  padding: 80px;
-  background-color: rgba(237, 225, 244, 0.929);
-}
-
-.Disable-container {
-  margin: 0 auto;
-  max-height: 250px;
-  padding: 20px;
-  max-width: 1200px;
-  min-height: 400px;
-  max-height: fit-content;
-}
-
-.DisableCard {
-  border: 1px solid #7b7979;
-  background-color: rgb(220, 218, 223);
-  max-width: 300px;
-  border-radius: 20px;
-  padding: 10px;
-  overflow: visible;
-  box-shadow: 0 0 30px rgba(0, 0, 0, 0.1);
-}
-
-.imgContainer {
-  display: flex;
-}
-
-.disableCard-img {
-  width: 250px;
-  height: 250px;
-  object-fit: cover;
-  border-radius: 20px;
-  -webkit-border-radius: 20px;
-  -moz-border-radius: 20px;
-  -ms-border-radius: 20px;
-  -o-border-radius: 20px;
-}
-
-.disableCard-title {
-  display: flex;
-  text-align: left;
-  justify-content: left;
-  padding: 10px;
-  font-weight: 800;
-}
-
-
-.disableCard-button {
-  display: flex;
-  align-items: right;
-  justify-content: right;
-  text-orientation:sideways-right;
-  padding: 10px;
-  border-radius: 10px;
-  -webkit-border-radius: 10px;
-  -moz-border-radius: 10px;
-  -ms-border-radius: 10px;
-  -o-border-radius: 10px;
-  background-color: #2b124c;
-  color: white;
-  font-weight: 500;
-  border: none;
-  cursor: pointer;
-}
